# Flush `AsyncPoolSink` on the current thread if tearing down