Fixes #64.

I've figured out what happened with this bug. There are 2 things went wrong:

1. `crossbeam` uses thread-local internally, which is not supported in `atexit` callback. This is the reason of panic `use of std::thread::current() is not possible after the thread's local data has been destroyed`.

   Fixed it by directly flushing the sinks on the current thread if the program is tearing down.

2. The default logger just gets flushed in `atexit` callback, but not dropped. This results in the program not waiting for the asynchronous thread to finish but simply exiting after the task has been sent.

   ~Fixed it by changing the global variable with `Option<T>`, so that we could explicitly drop the default logger in `atexit` callback.~ Reverted, paragraph 3 addresses the problem.

3. **UPDATE:** Destroying the thread pool before the final flush is required as well, otherwise pending log tasks may be lost.

   Fixing this by ~adding a `RwLock` to `ThreadPool`~ and explicitly destroy the thread pool before the final flush, but needs to be benchmarked to see if the performance change is acceptable.

   **UPDATE:** Changed to `ArcSwapOption` for a better performance.

---

Related discussion #63 #65, CC @MrCoco5921 @Bestfast