@SpriteOvO 2024-03-22
## Benchmark results

On ArchLinux, commit e1f9a068b1c28dcd223ba121df7853f9dc263e65

### Before

```
[info] --------------------------------------------
[info] Messages     : 1000000
[info] Threads      : 10
[info] Queue        : 8192 slots
[info] Queue memory : 8192 x 120 = 960 KB
[info] Total iters  : 3
[info] --------------------------------------------
[info] 
[info] ********************************************
[info] Queue Overflow Policy: Block
[info] ********************************************
[info] Elapsed: 0.355563435 secs         2812437/sec
[info] Elapsed: 0.392709661 secs         2546410/sec
[info] Elapsed: 0.350509562 secs         2852989/sec
[info] 
[info] ********************************************
[info] Queue Overflow Policy: DropIncoming
[info] ********************************************
[info] Elapsed: 0.080866536 secs         12366054/sec
[info] Elapsed: 0.077687493 secs         12872084/sec
[info] Elapsed: 0.075060614 secs         13322566/sec
```

### After

```
[info] --------------------------------------------
[info] Messages     : 1000000
[info] Threads      : 10
[info] Queue        : 8192 slots
[info] Queue memory : 8192 x 120 = 960 KB
[info] Total iters  : 3
[info] --------------------------------------------
[info] 
[info] ********************************************
[info] Queue Overflow Policy: Block
[info] ********************************************
[info] Elapsed: 0.417248511 secs         2396653/sec
[info] Elapsed: 0.434975067 secs         2298982/sec
[info] Elapsed: 0.406287435 secs         2461311/sec
[info] 
[info] ********************************************
[info] Queue Overflow Policy: DropIncoming
[info] ********************************************
[info] Elapsed: 0.123620255 secs         8089289/sec
[info] Elapsed: 0.104598975 secs         9560323/sec
[info] Elapsed: 0.107852031 secs         9271962/sec
```

Guess the main reason for the worse is the new `RwLock` added to `ThreadPool`, I will try to see if there is a solution with better performance.