@SpriteOvO 2024-03-22
If we use `ArcSwapOption` instead of `RwLock`, we could get a nice performance - almost no impact.

```
[info] --------------------------------------------
[info] Messages     : 1000000
[info] Threads      : 10
[info] Queue        : 8192 slots
[info] Queue memory : 8192 x 120 = 960 KB
[info] Total iters  : 3
[info] --------------------------------------------
[info] 
[info] ********************************************
[info] Queue Overflow Policy: Block
[info] ********************************************
[info] Elapsed: 0.339878523 secs         2942227/sec
[info] Elapsed: 0.388267234 secs         2575545/sec
[info] Elapsed: 0.357158248 secs         2799879/sec
[info] 
[info] ********************************************
[info] Queue Overflow Policy: DropIncoming
[info] ********************************************
[info] Elapsed: 0.071169735 secs         14050916/sec
[info] Elapsed: 0.075486378 secs         13247423/sec
[info] Elapsed: 0.074043964 secs         13505489/sec
```