@SpriteOvO 2024-03-22
Reverted commit `Drop the default logger in atexit callback`. We have explicitly destroyed the thread pool, so no longer need the drop function to wait for tasks to complete. Let's minimize the changes, keep the rule in Rust that global variables don't get dropped.